benc: Move to own pkg
