--- structdef-benc.go
@@ -1,10 +1,0 @@
-package goserbench
-
-type BENC struct {
-	Name     string
-	BirthDay int64
-	Phone    string
-	Siblings int32
-	Spouse   bool
-	Money    float64
-}
